requirements.txt
Flask
tensorflow
numpy
pandas
scikit-learn
requests
python-whois
gunicorn